보너스 끝

+++ philo_bonus/src/main.c
@@ -6,7 +6,7 @@
 /*   By: hkim2 <hkim2@student.42seoul.kr>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/27 00:40:55 by hkim2             #+#    #+#             */
-/*   Updated: 2022/04/28 18:43:45 by hkim2            ###   ########.fr       */
+/*   Updated: 2022/04/28 20:57:04 by hkim2            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,9 +17,12 @@ int	run_process(t_info *info)
 	int			idx;
 	pthread_t	must_thread;
 
-	if (pthread_create(&must_thread, NULL, check_must_eat_routine, (void *)info))
-		return (FAILURE);
-	pthread_detach(must_thread);
+	if (info->must_eat > 0)
+	{
+		if (pthread_create(&must_thread, NULL, check_must_eat_routine, (void *)info))
+			return (FAILURE);
+		pthread_detach(must_thread);
+	}
 	idx = -1;
 	while (++idx < info->num_of_philo)
 	{
@@ -29,34 +32,32 @@ int	run_process(t_info *info)
 			routine(&info->philos[idx]);
 			exit(SUCCESS);
 		}
-		usleep(100);
+		usleep(10);
 	}
 	return (SUCCESS);
 }
 
-void	philo_free(t_info *info)
+int	philo_free(t_info *info)
 {
 	int	idx;
 
 	idx = 0;
-	while (idx < info->num_of_philo)
+	if (info->philos)
 	{
-		kill(info->philos[idx].pid, SIGKILL);
-		idx++;
+		while (idx < info->num_of_philo)
+		{
+			kill(info->philos[idx].pid, SIGKILL);
+			idx++;
+		}
 	}
-	//sem_unlink(info->sem_fork);
-	//sem_unlink(info->sem_die);
-	//sem_unlink(info->sem_eat);
-	//sem_unlink(info->sem_print);
-	//sem_unlink(info->sem_stop);
-	//sem_unlink(info->sem_must_eat);
-	//sem_unlink(info->sem_fork);
-	//sem_unlink(info->sem_die);
-	//sem_unlink(info->sem_eat);
-	//sem_unlink(info->sem_print);
-	//sem_unlink(info->sem_stop);
-	//sem_unlink(info->sem_must_eat);
+	sem_unlink(info->fork);
+	sem_unlink(info->die);
+	sem_unlink(info->eat);
+	sem_unlink(info->print);
+	sem_unlink(info->stop);
+	sem_unlink(info->must);
 	free(info->philos);
+	return (SUCCESS);
 }
 
 
@@ -67,17 +68,15 @@ int	main(int argc, char **argv)
 	if (parse_argv(argc, argv, &info))
 		return (FAILURE);
 	if (init_philo(&info))
-		return (FAILURE);
+		return (philo_free(&info));
 	info.start_time = get_time();
 	sem_wait(info.sem_stop);
 	if (run_process(&info))
-		return (FAILURE);
+		return (philo_free(&info));
 	sem_wait(info.sem_stop);
 	philo_free(&info);
-	while(1)
-	{
-		NULL;
-	}
+	usleep(1000);
 	printf("finish\n");
+	
 	return (SUCCESS);
 }
